+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Calendar.js - Change Log:
 
+### Version 1.4.5:
+- Added new setting "minimumDatePickerDate", which states the minimum date to use while in DatePicker mode.
+- Added new setting "maximumDatePickerDate", which states the maximum date to use while in DatePicker mode.
+- The date selected in DatePicker mode is now highlighted in the calendar display.
+- Fixed a fault that caused the calendar to close when clicking in the days header (in DatePicker mode).
+- Fixed a fault that caused the Search dialog to move uncontrollably when a context menu is opened on the title bar.
+
 ### Version 1.4.4:
 - Minor comment updates to headers.
 - Added parameter "updateEvents" support to the public function addHolidays().